fix failed ci

+++ src/board_controller/aavaa/aavaa_v3.cpp
@@ -506,7 +506,7 @@ void AAVAAv3::read_data (
         Incoming_BLE_Data_Buffer.insert (Incoming_BLE_Data_Buffer.end (), data, data + size);
     }
 
-    while (Incoming_BLE_Data_Buffer.size () >= SIZE_OF_DATA_FRAME)
+    while ((int)Incoming_BLE_Data_Buffer.size () >= SIZE_OF_DATA_FRAME)
     {
         if (Incoming_BLE_Data_Buffer[0] != START_BYTE)
         {
@@ -558,8 +558,7 @@ void AAVAAv3::read_data (
         package[board_descr["default"]["battery_channel"].get<int> ()] = (double)data_frame[40];
 
         // imu status byte
-        package[board_descr["default"]["other_channels"][0].get<int> ()] =
-            (double)data_frame[41];
+        package[board_descr["default"]["other_channels"][0].get<int> ()] = (double)data_frame[41];
 
         // timestamp
         try
